requirements.txt
psycopg2>=2.5.2
Django>=1.6.4
django-userena>=1.2.1
Scrapy>=0.22.2
requests>=2.0.0
django-json-field>=0.5.5
pytidylib>=0.1.2
django-crispy-forms>=1.3.2
django-endless-pagination>=2.0
PyEnchant>=1.6.5
